Initial MapDB off-heap support; map, queue and multimap
TODO: Multimap is probably quite inefficient as the HZ code assumes in memory collections

Just store the values off heap, and use a long as a pointer into MapDB

Track the record ids in each MultiMapWrapper; use built in LONG serializer

Include hppc and mapdb in the "all" jar

Use the built in store implementation to provide off-heap queue storage by default

Fix multimap support by using the some pointer-to-off-heap pattern.

BTree won't work because not only are the keys huge, but MapDB does not compact the tree when nodes are deleted, resulting in a memory leak.

Look at queue migration.

Remove unused dependency

Fix a memory leak where we weren't killing items in the data map

Add in a better benchmark / resiliancy test.
If a QueueContainer is replaced, make sure it is destroyed.

Force going through a specific address

Conflicts:
	hazelcast-all/pom.xml
	hazelcast/src/main/java/com/hazelcast/map/PartitionContainer.java
	hazelcast/src/main/java/com/hazelcast/map/record/ObjectRecordFactory.java
	hazelcast/src/main/java/com/hazelcast/queue/QueueService.java
	hazelcast/src/test/java/com/hazelcast/map/SizeEstimatorTest.java